feat: Add UserId to WorkoutDto and update WorkoutMapper for mapping

+++ replog-shared/Models/Sync/LogSync/UpdateLogSyncModel.cs
@@ -6,4 +6,4 @@ public record UpdateLogSyncModel(
     string MuscleGroupId,
     string ExerciseId,
     int NumberReps,
-    int MaxWeight);
+    double MaxWeight);
